+++ LocalPackages/App/Sources/App/SendModule/Modules/SendV3/SendV3View.swift
@@ -57,55 +57,11 @@ final class SendV3View: UIView {
     
     addSubview(scrollView)
     scrollView.addSubview(stackView)
-    
-    let liquidStakingView = StakingView(model: .init(
-      title: "Liquid Staking",
-      providers: [
-        .init(
-          iconURL: iconURL,
-          title: "Tonstakers",
-          subtitle: "Minimum deposit 1 TON.\nAPY ≈ 5.01%",
-          badge: "MAX APY",
-          isSelected: true
-        ),
-        .init(
-          iconURL: iconURL,
-          title: "Bemo",
-          subtitle: "Minimum deposit 1 TON.\nAPY ≈ 4.01%",
-          badge: nil,
-          isSelected: false
-        ),
-      ],
-      style: .liquid
-    ))
-    stackView.addArrangedSubview(liquidStakingView)
 
-    let otherStakingView = StakingView(model: .init(
-      title: "Other",
-      providers: [
-        .init(
-          iconURL: iconURL,
-          title: "Tonstakers",
-          subtitle: "Minimum deposit 1 TON.\nAPY ≈ 5.01%",
-          badge: "MAX APY",
-          isSelected: true
-        ),
-        .init(
-          iconURL: iconURL,
-          title: "Bemo",
-          subtitle: "Minimum deposit 1 TON.\nAPY ≈ 4.01%",
-          badge: nil,
-          isSelected: false
-        ),
-      ],
-      style: .other
-    ))
-    stackView.addArrangedSubview(otherStakingView)
-
-//    stackView.addArrangedSubview(recipientTextField)
-//    stackView.addArrangedSubview(amountInputView)
-//    stackView.addArrangedSubview(commentInputView)
-//    stackView.addArrangedSubview(continueButton)
+    stackView.addArrangedSubview(recipientTextField)
+    stackView.addArrangedSubview(amountInputView)
+    stackView.addArrangedSubview(commentInputView)
+    stackView.addArrangedSubview(continueButton)
     
     scrollView.snp.makeConstraints { make in
       make.edges.equalTo(self)
