+++ LocalPackages/App/Sources/App/StakeModule/Modules/Stake/StakeViewController.swift
@@ -142,7 +142,7 @@ private extension StakeViewController {
   
   @objc
   func didTapInfoButton() {
-    // TODO: Add action
+    viewModel.didTapToInfo()
   }
   
 }
